+++ database/migrations/2022_05_15_105434_create_gas_bills_table.php
@@ -16,7 +16,7 @@ class CreateGasBillsTable extends Migration
         Schema::create('gas_bills', function (Blueprint $table) {
             $table->id();
             $table->bigInteger('bill_id');
-            $table->bigInteger('cost');
+            $table->float('cost');
             $table->timestamps();
         });
     }
