+++ database/migrations/2022_05_15_105540_create_electricity_bills_table.php
@@ -16,7 +16,7 @@ class CreateElectricityBillsTable extends Migration
         Schema::create('electricity_bills', function (Blueprint $table) {
             $table->id();
             $table->bigInteger('bill_id');
-            $table->bigInteger('cost');
+            $table->float('cost');
             $table->timestamps();
         });
     }
